refactor: update NavBar goback param
