+++ packages/experience/src/hooks/use-email-blocked-error-handler.ts
@@ -1,7 +1,7 @@
 import { type RequestErrorBody } from '@logto/schemas';
 import { useCallback, useMemo } from 'react';
 
-import useNavigateWithPreservedSearchParams from '@/shared/hooks/use-navigate-with-preserved-search-params';
+import useNavigateWithPreservedSearchParams from '@/hooks/use-navigate-with-preserved-search-params';
 
 import { usePromiseConfirmModal } from './use-confirm-modal';
 import { type ErrorHandlers } from './use-error-handler';
