+++ packages/experience/src/hooks/use-send-mfa-verification-code.ts
@@ -5,7 +5,7 @@ import UserInteractionContext from '@/Providers/UserInteractionContextProvider/U
 import { sendMfaVerificationCode } from '@/apis/experience';
 import useApi from '@/hooks/use-api';
 import useErrorHandler from '@/hooks/use-error-handler';
-import useNavigateWithPreservedSearchParams from '@/shared/hooks/use-navigate-with-preserved-search-params';
+import useNavigateWithPreservedSearchParams from '@/hooks/use-navigate-with-preserved-search-params';
 import { type VerificationCodeIdentifier } from '@/types';
 import { type MfaFlowState } from '@/types/guard';
 import { codeVerificationTypeMap } from '@/utils/sign-in-experience';
