+++ packages/experience/src/hooks/use-send-verification-code.ts
@@ -9,7 +9,7 @@ import UserInteractionContext from '@/Providers/UserInteractionContextProvider/U
 import { sendVerificationCodeApi } from '@/apis/utils';
 import useApi from '@/hooks/use-api';
 import useErrorHandler from '@/hooks/use-error-handler';
-import useNavigateWithPreservedSearchParams from '@/shared/hooks/use-navigate-with-preserved-search-params';
+import useNavigateWithPreservedSearchParams from '@/hooks/use-navigate-with-preserved-search-params';
 import {
   UserFlow,
   type ContinueFlowInteractionEvent,
