+++ packages/experience/src/pages/ErrorPage/index.tsx
@@ -6,6 +6,7 @@ import StaticPageLayout from '@/Layout/StaticPageLayout';
 import PageContext from '@/Providers/PageContextProvider/PageContext';
 import EmptyStateDark from '@/assets/icons/empty-state-dark.svg?react';
 import EmptyState from '@/assets/icons/empty-state.svg?react';
+import useNavigateWithPreservedSearchParams from '@/hooks/use-navigate-with-preserved-search-params';
 import DynamicT from '@/shared/components/DynamicT';
 import NavBar from '@/shared/components/NavBar';
 import PageMeta from '@/shared/components/PageMeta';
@@ -27,12 +28,20 @@ const ErrorPage = ({
   isNavbarHidden,
 }: Props) => {
   const { theme } = useContext(PageContext);
+  const navigate = useNavigateWithPreservedSearchParams();
   const errorMessage = Boolean(rawMessage ?? message);
 
   return (
     <StaticPageLayout>
       <PageMeta titleKey={title} />
-      {history.length > 1 && <NavBar isHidden={isNavbarHidden} />}
+      {history.length > 1 && (
+        <NavBar
+          isHidden={isNavbarHidden}
+          onBack={() => {
+            navigate(-1);
+          }}
+        />
+      )}
       <div className={styles.container}>
         {theme === Theme.Light ? <EmptyState /> : <EmptyStateDark />}
         <div className={styles.title}>
